**Flite 1.1 upgrade - FeatureProcessor upgrades **
AWB did some refactoring of the Flite Feature processors, replacing
duplicated code with a set of functions that are called.  I've
paralled this refactoring in FreeTTS.

Flite now seems to try to eliminate the need for any static or global
data (probably as part of the Wince support), many flite changes in
the FeatureProcessor area have to do with hanging the feature lists
off of the voice now instead of in a global area. Since we were
already doing the same in FreeTTS we did not have to do anything about
that.

A number of new FeatureProcessors have been added:

WordPunc
SegCodaFric
SegOnsetFric
SegCodeGlide
SegOnsetGlide
SegCodeNasal
SegOnsetNasal
SegmentDuration

Also, note that several of the bug fixes that we applied to Flite in
this set of routines have NOT be incorporated back into the main
sources by AWB and had to be re-fixed.

============
Test Status
============
At this point all of the JUNIT tests still pass, but some of the
regression tests are failing. This is to be expected, since the
regression tests are testing against the output of Flite 1.0. Until
the tests are upgraded to test against the output of Flite 1.1, they
will continue to fail.