Fixed a bug in LPCResult. Description of the bug: LPCResult writes out audio samples to the AudioPlayer in batches of size 1024 (or any specified size). If there happens to be exactly 1024 samples in the very last batch, the AudioPlayer.end() method will not be called. However, audio data for a single Utterance should be grouped between AudioPlayer.begin() and AudioPlayer.end() calls, so we have a bug.
This bug is the cause for the "From:" input text hanging the ClientServer
demo, as reported in the FreeTTS help message board.
The Server is relying on FreeTTS calling AudioPlayer.end() to tell
the Client that the Utterance has ended. The last batch of audio of "From:"
just happens to be 1024, so AudioPlayer.end() was never called.

It is now fixed such that as long as the AudioPlayer is working, the
end() method will be called, regardless of the size of the last batch of
audio.