Fixed a bug in Voice.java. The mods to this fix was actually in the previous commit. Refer to that commit diffs to see where the code was changed.
Description of the bug:

In class Voice.FreeTTSSpeakableTokenizer.Iterator, the method

public Object next()

which returns the next Utterance of the FreeTTSSpeakable,
will incorrectly call Utterance.setLast(true), under a particular situation.
An example of the situation is when the input is:

To: Janet

With the bug, since ":" signifies a break, the method next() will return
an Utterance with only "To" as the word. However, since the Tokenizer
has already read "Janet" (saved as "savedToken"), there is no more
text to read, so the next() method thinks that this
Utterance (with "To" only) is the last Utterance.

The fix is to set an Utterance as the last one only if:

!tok.hasMoreTokens() && savedToken == null

This fixes the "To: Janet" bug reported in the FreeTTS help message board.